remove mmgr and macros that redefine malloc/new/free (see http://www.wildfiregames.com/forum/index.php?showtopic=11450&hl= )
clean up debug module
. no longer include platform-dependent header (-> less rebuilds)
. DISPLAY_ERROR -> DEBUG_DISPLAY_ERROR
. parts of config.h that don't affect all files moved to config.2 (-> fewer full rebuilds)
. remove creaky symbol cache (no longer needed for mmgr)
. remove TLS thread naming stuff (can use debugger's thread window instead; no need for platform independence there)

wdbg: remove thread suspension and breakpoint APIs (not needed)

acpi: fix: u64 -> uintptr_t

wutil: fix WinScopedLock, use that instead of direct lock() functions

misc:
. get rid of SAFE_STRCPY, replace with strcpy_s
. remove _getcwd (shouldn't be used)

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@5563 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/sysdep/acpi.cpp
@@ -51,7 +51,7 @@ static inline void* CallWithSafetyBlanket(UnsafeFunction func, PCV_u8 mem, size_
 #endif
 }
 
-static void* TransactPhysicalMemory(u64 physicalAddress, size_t numBytes, UnsafeFunction func, void* arg = 0)
+static void* TransactPhysicalMemory(uintptr_t physicalAddress, size_t numBytes, UnsafeFunction func, void* arg = 0)
 {
 	PCV_u8 mem = (PCV_u8)mahaf_MapPhysicalMemory(physicalAddress, numBytes);
 	if(!mem)
@@ -168,7 +168,7 @@ static inline void* UnsafeAllocateCopyOfTable(PCV_u8 mem, size_t numBytes, void*
 
 // caller is responsible for verifying the table is valid and using
 // DeallocateTable to free it.
-static const AcpiTable* AllocateCopyOfTable(u64 physicalAddress)
+static const AcpiTable* AllocateCopyOfTable(uintptr_t physicalAddress)
 {
 	// ACPI table sizes are not known until they've been mapped. since that
 	// is slow, we don't always want to do it twice. the solution is to map
@@ -229,7 +229,7 @@ static bool VerifyTable(const AcpiTable* table, const char* signature = 0)
 }
 
 
-static const AcpiTable* GetTable(u64 physicalAddress, const char* signature = 0)
+static const AcpiTable* GetTable(uintptr_t physicalAddress, const char* signature = 0)
 {
 	const AcpiTable* table = AllocateCopyOfTable(physicalAddress);
 	if(VerifyTable(table, signature))
